AstroForecastFunction/__init__.py

import logging
import azure.functions as func
from datetime import datetime
import pandas as pd

def main(mytimer: func.TimerRequest) -> None:
    utc_timestamp = datetime.utcnow().isoformat()
    logging.info(f"Astro Forecast Function ran at {utc_timestamp}")

    # Sample logic placeholder
    df = pd.DataFrame({"Date": [utc_timestamp], "Signal": ["Sample"]})
    df.to_csv("/tmp/astro_forecast.csv", index=False)
    logging.info("Sample forecast CSV saved.")
