requirements.txt

azure-functions
pandas
